ingestor/requirements.txt
yfinance
openai==1.14
azure-search-documents
azure-identity
redis
python-dotenv
azure-ai-inference